backend/requirements.txt
# Django + DRF
Django==4.0.2
djangorestframework
djangorestframework-simplejwt

# 引数設定・データベース
python-dotenv
psycopg2-binary

# AI関連
langchain-huggingface
langchain-community
faiss-cpu
openai
